+++ db/importer.js
@@ -1,5 +1,5 @@
 /**
- *  ledgerImporter uses the stellard API to import
+ *  ledgerImporter uses the paysharesd API to import
  *  ledgers into a CouchDB instance
  *
  *  Available command line options:
@@ -21,8 +21,8 @@ function ledgerImporter () {
     moment   = require('moment'),
     _        = require('lodash'),
     diff     = require('deep-diff'),
-    stellar   = require('stellar-lib'),
-    Ledger   = require('../node_modules/stellar-lib/src/js/ripple/ledger').Ledger,
+    payshares   = require('payshares-lib'),
+    Ledger   = require('../node_modules/payshares-lib/src/js/ripple/ledger').Ledger,
     options  = {};
   
   var env    = process.env.NODE_ENV || "development",
@@ -91,7 +91,7 @@ function ledgerImporter () {
     //importer from the last point where its correct.
     
     
-    //get the last closed ledger from stellard
+    //get the last closed ledger from paysharesd
     //work back from there
     
     if (!opts.minLedgerIndex) {
@@ -202,9 +202,9 @@ function ledgerImporter () {
 
 
  /*
-  *  importLive: import live data from stellard
+  *  importLive: import live data from paysharesd
   * 
-  *  first, get the latest ledger from stellard and save it to the database.   
+  *  first, get the latest ledger from paysharesd and save it to the database.   
   *  we do this so that we can check every subsequent batch of live ledgers
   *  against the ledger chain stored in the database.  
   * 
@@ -469,7 +469,7 @@ function ledgerImporter () {
   
  /**
   *  getLedgerBatch starts from a specified ledger or the most recently
-  *  closed one and uses the stellard API to get the batch of ledgers
+  *  closed one and uses the paysharesd API to get the batch of ledgers
   *  one by one, walking the ledger hash chain backwards until it reaches the minLedger
   *
   *  Available options:
@@ -496,7 +496,7 @@ function ledgerImporter () {
       opts.results = [];
     }
   
-    // get ledger from stellard API
+    // get ledger from paysharesd API
     getLedger({
       identifier: opts.lastLedger
     }, function(err, ledger){
@@ -560,7 +560,7 @@ function ledgerImporter () {
   
   
  /**
-  *  getLedger uses the stellard API to get the ledger
+  *  getLedger uses the paysharesd API to get the ledger
   *  corresponding to the given identifier, or the last 
   *  closed ledger if identifier is null
   *
@@ -598,7 +598,7 @@ function ledgerImporter () {
   
     // store server statuses (reset for each ledger identifier)
     if (!servers) {
-      servers = _.map(config.stellards, function(serv){
+      servers = _.map(config.paysharesds, function(serv){
         return {
           server: serv,
           attempt: 0
@@ -612,7 +612,7 @@ function ledgerImporter () {
     if (serverEntry.attempt >= 2) {
       callback('ledger ' + 
         (reqData.params[0].ledger_index || reqData.params[0].ledger_hash) + 
-        ' not available from any of the stellards');
+        ' not available from any of the paysharesds');
       return;
     }
   
@@ -647,7 +647,7 @@ function ledgerImporter () {
   
       // check if the server returned a buffer/string instead of json
       if (typeof res.body === 'string' || res.body.constructor.name === 'Buffer') {
-        // console.log('stellard returned a buffer instead of a JSON object for request: ' + 
+        // console.log('paysharesd returned a buffer instead of a JSON object for request: ' + 
         //   JSON.stringify(reqData.params[0]) + '. Trying again...');
         // servers[server] = 'tryAgain';
   
@@ -707,7 +707,7 @@ function ledgerImporter () {
       // check for malformed ledger
       if (!ledger || !ledger.ledger_index || !ledger.ledger_hash) {
         winston.info('got malformed ledger from ' + 
-          (server === 'http://0.0.0.0:51234' ? 'http://ct.stellar.com:51234' : server) + ': ' + 
+          (server === 'http://0.0.0.0:51234' ? 'http://ct.payshares.com:51234' : server) + ': ' + 
           JSON.stringify(ledger));
   
         _.find(servers, function(serv){ return serv.server === server; }).attempt++;
@@ -724,7 +724,7 @@ function ledgerImporter () {
       }
   
       // keep track of which server ledgers came from
-      ledger.server = (server === 'http://0.0.0.0:51234' ? 'http://ct.stellar.com:51234' : server);
+      ledger.server = (server === 'http://0.0.0.0:51234' ? 'http://ct.payshares.com:51234' : server);
   
       // check that transactions hash to the expected value
       var ledgerJsonTxHash;
@@ -770,10 +770,10 @@ function ledgerImporter () {
   function formatRemoteLedger(ledger) {
   
     ledger.close_time_rpepoch = ledger.close_time;
-    ledger.close_time_timestamp = stellar.utils.toTimestamp(ledger.close_time);
-    ledger.close_time_human = moment(stellar.utils.toTimestamp(ledger.close_time))
+    ledger.close_time_timestamp = payshares.utils.toTimestamp(ledger.close_time);
+    ledger.close_time_human = moment(payshares.utils.toTimestamp(ledger.close_time))
       .utc().format("YYYY-MM-DD HH:mm:ss Z");
-    ledger.from_stellard_api = true;
+    ledger.from_paysharesd_api = true;
   
     delete ledger.close_time;
     delete ledger.hash;
@@ -804,7 +804,7 @@ function ledgerImporter () {
         var fields = node.FinalFields || node.NewFields;
   
         if (typeof fields.BookDirectory === "string") {
-          node.exchange_rate = stellar.Amount.from_quality(fields.BookDirectory).to_json().value;
+          node.exchange_rate = payshares.Amount.from_quality(fields.BookDirectory).to_json().value;
         }
   
       });