+++ release-notes.md
@@ -2,14 +2,14 @@
 
 ## Development
 
-[Commits](https://github.com/stellar/stellar-data-api/commits/master)
+[Commits](https://github.com/payshares/payshares-data-api/commits/master)
 
 ##December 2014
 
 ###New Features and Changes
 
- - [2](https://github.com/stellar/stellar-data-api/pull/2) Add API for STR and USD for bitcoin charts  
- - [1](https://github.com/stellar/stellar-data-api/pull/1) Ports and stellarizes ripple charts for Stellar 
+ - [2](https://github.com/payshares/payshares-data-api/pull/2) Add API for XPR and USD for bitcoin charts  
+ - [1](https://github.com/payshares/payshares-data-api/pull/1) Ports and paysharesizes ripple charts for Payshares